refactor: split evaluate kernel behind stable facade
