+++ crates/assay-core/src/mcp/tool_call_handler/mod.rs
@@ -6,6 +6,7 @@
 
 mod emit;
 mod evaluate;
+mod evaluate_next;
 mod types;
 
 pub use types::{HandleResult, ToolCallHandler, ToolCallHandlerConfig};
